feat(models): Code Factorization
The repos was (is still...) full of code duplication. This commit deletes parts of it. However, while improving the code, I also made modification to improve the reproducibility of the experiments.
Namely, before each model training, seeds are reinitialised. Also, SecureBERT loading and initialization is now deterministic, the same embeddings are providing on different runs (not the case before).
This will marginally change the performances of the models, however the trends remains the same.

In more details, the following code parts were factorized:
- CountVectorizer, Li and SecureBERT models all use the compute_metrics_generic function;
- We provide for each model type, specific preprocessing functions as needs, same for decision scoring functions;
- SecureBERT models now inherite from a base class to reduce duplicate code.